ЛР 2 по ЧМ

+++ sem4/met/lab1/1_2.py
@@ -7,7 +7,7 @@ y = sympy.Symbol('y')
 expr = (2/(x**2) + 3/(y**3))*(x+3*y)
 
 sympy.simplify(expr)
-print(float(expr.subs(x, -2.01).subs(y, sympy.sqrt(5))))
+print(expr.subs(x, -2.01).subs(y, sympy.sqrt(5)))
 
 print("2 задание: ")
 print(sympy.diff(expr, x))
